fixed react bug by adding unique keys
