+++ src/RowHeader.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import FilterDropdown from './FilterDropdown';
-import FilterDropdownNative from './FilterDropdownNative';
+import FilterDropdown from './FilterDropdown'
 
 type RowHeaderProps = {
   api: string,
@@ -13,16 +12,13 @@ type RowHeaderProps = {
 }
 
 export default function RowHeader(props:RowHeaderProps):React.JSX.Element {
-
   return (
       <div className='rowHeader'>
         <div className='headerName-api'>{props.api}</div>
         <div className='headerName'>{props.location}</div>
         <div className='headerName-status'>
-          <div>{props.status}</div>
-          <div className="filter-dropdown">
-            < FilterDropdownNative filters={props.filters} filter={props.filter}/>
-          </div>
+          {props.status}
+          < FilterDropdown filters={props.filters} filter={props.filter}/>
         </div>
         <div className='headerName-stable'>{props.stable}</div>
         <div className='headerName-notes'>{props.notes}</div>
